Create gui.cpp
Update main.cpp

Update makefile

Update terminal.cpp

Update terminalgui.hpp

Update serial.cpp
