+++ gui/terminal.cpp
@@ -3,97 +3,6 @@
 #include <fstream>
 #include <stdio.h>
 
-void uiInit(HWND* appwnd, uiElements* ui, appData* data) {
-
-	//	serial speeds dropdown
-	//	it isn't gonna be updated coz speeds are hardcoded to the library
-	//	so render it only once and then just use as intended
-	ui->comboSpeed = CreateWindowA(WC_COMBOBOXA, NULL, WS_CHILD | WS_VISIBLE | CBS_DROPDOWNLIST | CBS_SIMPLE | WS_VSCROLL, 500, 8, 120, 200, *appwnd, (HMENU)GUI_COMBO_SPEED, NULL, NULL);  
-	{
-		std::vector <std::string> temp;
-
-		for (auto item : data->speeds) {
-			temp.push_back(std::to_string(item) + " baud");
-		}
-
-		//	select 9600, bc it's the default
-		for (size_t i = 1; i < data->speeds.size(); i++) {
-			if (data->speeds.at(i) == SIO_DEFAULT_SPEED) {
-				data->sel_speed = i;
-				break;
-			}
-		}
-
-		dropdown(&ui->comboSpeed, &temp, data->sel_speed, false);
-	}
-
-	//	port selector
-	//	the items are gonna be assigned by the update timer callback
-	//	so gonna leave it empty for now
-	ui->comboPort = CreateWindowA(WC_COMBOBOXA, NULL, WS_CHILD | WS_VISIBLE | CBS_DROPDOWNLIST | CBS_SIMPLE | WS_VSCROLL, 420, 8, 80, 200, *appwnd, (HMENU)GUI_COMBO_PORT, NULL, NULL);
-
-	//	line ending selector
-	ui->comboLineEnding = CreateWindowA(WC_COMBOBOXA, NULL, WS_CHILD | WS_VISIBLE | CBS_DROPDOWNLIST | CBS_SIMPLE | WS_VSCROLL, 340, 8, 80, 200, *appwnd, (HMENU)GUI_COMBO_LINE, NULL, NULL);
-	{
-		std::vector <std::string> temp;
-		for (auto item : data->endlines) temp.push_back(item.title);
-		dropdown(&ui->comboLineEnding, &temp, data->sel_endline, false);
-	}
-
-	//	terminal window itself
-	ui->terminal = CreateWindowA(WC_EDITA, NULL, WS_CHILD | WS_VISIBLE | WS_BORDER | WS_VSCROLL | ES_MULTILINE | ES_READONLY, 0, 40, 630, 300, *appwnd, (HMENU)GUI_LOGWIN, NULL, NULL);	
-		
-	//	input	
-	ui->command = CreateWindowA(WC_EDITA, NULL, WS_VISIBLE | WS_CHILD | ES_LEFT | WS_BORDER, 10, 350, 568, 24, *appwnd, (HMENU)GUI_COMPROM, NULL, NULL);
-	
-	//	buttons
-	ui->btnSend = CreateWindowA(WC_BUTTONA, "Send", WS_VISIBLE | WS_CHILD | BS_BITMAP, 588, 350, 25, 25, *appwnd, (HMENU)GUI_BTN_SEND, NULL, NULL);
-	{
-		HBITMAP	image_send = LoadBitmap(GetModuleHandle(0), MAKEINTRESOURCE(ICON_BUTTON_SEND));
-			if (image_send) SendMessage(ui->btnSend, BM_SETIMAGE, (WPARAM)IMAGE_BITMAP, (LPARAM)image_send);
-			else printf("didn't load: %i\n", GetLastError());
-	}
-		
-	//	checkboxes
-	ui->timestamps = CreateWindowA(WC_BUTTONA, "Show timestamps", WS_VISIBLE | WS_CHILD | BS_VCENTER | BS_AUTOCHECKBOX, 10, 10, 110, 16, *appwnd, (HMENU)CHECKBOX_TIMESTAMP, NULL, NULL);
-	SendMessageW(ui->timestamps, BM_SETCHECK, BST_CHECKED, 0);
-	
-	ui->echoCommands = CreateWindowA(WC_BUTTONA, "Echo commands", WS_VISIBLE | WS_CHILD | BS_VCENTER | BS_AUTOCHECKBOX, 125, 10, 105, 16, *appwnd, (HMENU)CHECKBOX_ECHOCMD, NULL, NULL);
-	SendMessageW(ui->echoCommands, BM_SETCHECK, BST_CHECKED, 0);
-
-	//	status bar
-	ui->statusbar = CreateWindowA(WC_STATICA, "Starting up...", WS_VISIBLE | WS_CHILD | SS_LEFT, 5, 410, 200, 16, *appwnd, (HMENU)GUI_STATUSBAR, NULL, NULL);
-
-	//	set font
-	for (size_t i = 0; i <= 1000; i++) {
-		if (!GetDlgItem(*appwnd, i)) continue;
-		SendDlgItemMessage(*appwnd, i, WM_SETFONT, (WPARAM)GetStockObject(DEFAULT_GUI_FONT), MAKELPARAM(1, 0));
-	}
-}
-
-void dropdown(HWND* combo, std::vector <std::string>* items, size_t focus, bool erase) {
-	
-	if (erase) {
-		size_t contlen = SendMessage(*combo, CB_GETCOUNT, 0, 0);
-		for (size_t i = 0; i < contlen; i++)
-			SendMessage(*combo, CB_DELETESTRING, 0, 0);
-	}
-
-	for (auto entry : *items)
-		SendMessage(*combo, CB_ADDSTRING, 0, (LPARAM)entry.c_str());
-	
-	SendMessage(*combo, CB_SETCURSEL , focus, 0);
-}
-
-void displayAboutMessage() {
-
-	std::string msg = std::string(APP_NAME) + " v" + APP_VERSION +
-		"\nA serial port communication utility\n\n" +
-		VER_AUTHSTAMP + "\n" + APP_COPYRIGHT;
-		
-	MessageBoxA(NULL, msg.c_str(), "About...", 0);
-}
-
 void saveCommLog(HWND* appwnd, std::vector <std::string>* logdata) {
 
 	OPENFILENAMEA ofn;
@@ -230,32 +139,51 @@ void updateStatusBar(maddsua::serial* serial, uiElements* ui, appData* data) {
 
 void sendMessage(maddsua::serial* serial, uiElements* ui, appData* data) {
 
+	//	get input text
 	char userinput[1024];
 	if (!GetWindowTextA(ui->command, userinput, sizeof(userinput))) return;
 
+	//	check if there are ports
 	if (data->sel_port >= data->ports.size()) {
 		SetWindowTextA(ui->statusbar, "No active port selected!");
 		return;
 	}
 
+	//	echo to terminal
+	if (data->echoCommands) printComm(ui, data, userinput, false);
+
+	//	command history stuff
 	for (auto itr = data->history.begin(); itr != data->history.end(); itr++) {
 		if ((*itr) == userinput) {
 			data->history.erase(itr);
 			break;
 		}
 	}
-
 	data->history.push_back(userinput);
+	data->sel_history = data->history.size() ? (data->history.size() - 1) : 0;
 
+	//	write data to com port
 	auto port = data->ports.at(data->sel_port);
 	auto endline = data->endlines.at(data->sel_endline).bytes;
-
-	if (data->echoCommands) printComm(ui, data, userinput, false);
-
 	if (!serial->write(port, std::string(userinput) + endline)) {
 		SetWindowTextA(ui->statusbar, "Failed to send the message!");
 		return;
 	}
 
+	//	clear command prompt
 	SetWindowText(ui->command, NULL);
+}
+
+void historyRecall(uiElements* ui, appData* data, int direction) {
+
+	//	move history cursor
+	if ((data->sel_history + direction) < data->history.size())
+		data->sel_history += direction;
+
+	//	insert text
+	SetWindowTextA(ui->command, data->history.at(data->sel_history).c_str());
+
+	//	move text cursor to the end
+	auto textLen = SendMessage(ui->command, WM_GETTEXTLENGTH, 0, 0);
+	SendMessage(ui->command, EM_SETSEL, (WPARAM)textLen, (LPARAM)textLen);
 }
